+++ integration_tests/suite/base/test_schedules.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from hamcrest import (
@@ -467,7 +467,6 @@ def test_edit_all_parameters(schedule):
 
 
 @fixtures.schedule()
-@fixtures.meetme()
 @fixtures.ivr()
 @fixtures.group()
 @fixtures.outcall()