+++ wazo_confd/plugins/device/funckey.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import abc
@@ -155,7 +155,7 @@ class ConferenceConverter(FuncKeyConverter):
 
     def build(self, user, line, position, funckey):
         extension = self.extension_dao.get_by(
-            type='meetme', typeval=str(funckey.destination.conference_id)
+            type='conference', typeval=str(funckey.destination.conference_id)
         )
         return self.provd_funckey(line, position, funckey, extension.exten)